循环周期改為小數

+++ business/src/main/java/net/wanji/business/domain/TrafficFlow.java
@@ -22,8 +22,8 @@ public class TrafficFlow {
     //最小车流量
     Integer  minFlow;
 
-    //周期(min)
-    Integer  cycle;
+    //周期(h)
+    Double  cycle;
 
     //梯度
     Integer gradient;
